Refactor common package and move Skibidi into package
