+++ docs/README.md
@@ -1,4 +1,4 @@
-# Skibidi User Guide
+# skibidi.Skibidi User Guide
 
 // Update the title above to match the actual product name
