+++ src/main/java/commands/AddEventCommand.java
@@ -1,13 +1,12 @@
 package commands;
 
-import common.Command;
-import common.SkibidiException;
-import common.Ui;
+import skibidi.Command;
+import skibidi.SkibidiException;
+import skibidi.Ui;
 import storage.TaskStorage;
 import storage.Event;
 
 import java.io.IOException;
-import java.time.LocalDateTime;
 import java.time.format.DateTimeParseException;
 
 public class AddEventCommand extends Command {
