+++ src/main/java/skibidi/Command.java
@@ -1,8 +1,6 @@
-package common;
+package skibidi;
 
 import storage.TaskStorage;
-
-import java.io.IOException;
 
 public abstract class Command {
     public abstract boolean execute(Ui ui, TaskStorage storage) throws SkibidiException;
