+++ src/main/java/skibidi/SkibidiException.java
@@ -1,4 +1,4 @@
-package common;
+package skibidi;
 
 public class SkibidiException extends Exception {
     public SkibidiException(String message) {
